more dark/light mode, refactor thread card
